backend/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database
pymongo==4.9.1
motor==3.6.0  # Async MongoDB driver

# AI/ML
anthropic==0.39.0
openai==1.54.4  # Optional, if you want to use OpenAI
google-generativeai>=0.8.0  # Optional, if you want to use Gemini
langchain==0.3.7
langchain-community==0.3.5
sentence-transformers==3.3.1

# Vector Store
chromadb==0.5.20

# PDF Processing
pypdf2==3.0.1
pymupdf==1.24.13

# Document Processing
python-docx==1.1.0

# Authentication
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# Language Detection
lingua-language-detector>=2.0.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Testing (optional for now)
pytest==8.3.4
httpx==0.27.2